refactor: split telemetry queries and openrouter generations
